2021. 03. 10 Wed 18:50 C05 norm Chk

+++ C05/ex07/ft_find_next_prime.c
@@ -3,23 +3,23 @@
 /*                                                        :::      ::::::::   */
 /*   ft_find_next_prime.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: seungjle <seungjle@studnet.42seoul.>       +#+  +:+       +#+        */
+/*   By: seungjle <seungjle@student.42seoul.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/02 23:23:33 by seungjle          #+#    #+#             */
-/*   Updated: 2021/03/09 23:29:33 by seungjle         ###   ########.fr       */
+/*   Updated: 2021/03/10 18:48:20 by seungjle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_find_next_prime(int nb)
+int					ft_find_next_prime(int nb)
 {
-	int i;
+	unsigned int	i;
 
 	i = 2;
 	if (nb < 2)
 		return (ft_find_next_prime(nb + 1));
-	while (i * i < nb)
+	while (i * i < (unsigned int)nb)
 	{
-		if (nb % i == 0)
+		if ((unsigned int)nb % i == 0)
 			return (ft_find_next_prime(nb + 1));
 		i++;
 	}
